feat: add `zarf dev deploy` for quickly testing packages and restructure `zarf prepare` (#2170)
## Description

Adds a `zarf dev deploy` command that will create + deploy a local
package in one shot.

## Related Issue

Fixes #2169 
Fixes #2098 

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/zarf/blob/main/CONTRIBUTING.md#developer-workflow)
followed

---------
