+++ docs/2-the-zarf-cli/100-cli-commands/zarf.md
@@ -31,8 +31,8 @@ zarf COMMAND [flags]
 * [zarf completion](zarf_completion.md)	 - Generate the autocompletion script for the specified shell
 * [zarf connect](zarf_connect.md)	 - Accesses services or pods deployed in the cluster
 * [zarf destroy](zarf_destroy.md)	 - Tears down Zarf and removes its components from the environment
+* [zarf dev](zarf_dev.md)	 - Commands useful for developing packages
 * [zarf init](zarf_init.md)	 - Prepares a k8s cluster for the deployment of Zarf packages
 * [zarf package](zarf_package.md)	 - Zarf package commands for creating, deploying, and inspecting packages
-* [zarf prepare](zarf_prepare.md)	 - Tools to help prepare assets for packaging
 * [zarf tools](zarf_tools.md)	 - Collection of additional tools to make airgap easier
 * [zarf version](zarf_version.md)	 - Shows the version of the running Zarf binary
