+++ docs/2-the-zarf-cli/100-cli-commands/zarf_dev.md
@@ -1,20 +1,12 @@
-# zarf prepare lint
+# zarf dev
 <!-- Auto-generated by hack/gen-cli-docs.sh -->
 
-Verifies the package schema
-
-## Synopsis
-
-Verifies the package schema and warns the user if they have variables that won't be evaluated
-
-```
-zarf prepare lint [ DIRECTORY ] [flags]
-```
+Commands useful for developing packages
 
 ## Options
 
 ```
-  -h, --help   help for lint
+  -h, --help   help for dev
 ```
 
 ## Options inherited from parent commands
@@ -32,4 +24,11 @@ zarf prepare lint [ DIRECTORY ] [flags]
 
 ## SEE ALSO
 
-* [zarf prepare](zarf_prepare.md)	 - Tools to help prepare assets for packaging
+* [zarf](zarf.md)	 - DevSecOps for Airgap
+* [zarf dev deploy](zarf_dev_deploy.md)	 - [beta] Creates and deploys a Zarf package from a given directory
+* [zarf dev find-images](zarf_dev_find-images.md)	 - Evaluates components in a Zarf file to identify images specified in their helm charts and manifests
+* [zarf dev generate-config](zarf_dev_generate-config.md)	 - Generates a config file for Zarf
+* [zarf dev lint](zarf_dev_lint.md)	 - Verifies the package schema
+* [zarf dev patch-git](zarf_dev_patch-git.md)	 - Converts all .git URLs to the specified Zarf HOST and with the Zarf URL pattern in a given FILE.  NOTE:
+This should only be used for manifests that are not mutated by the Zarf Agent Mutating Webhook.
+* [zarf dev sha256sum](zarf_dev_sha256sum.md)	 - Generates a SHA256SUM for the given file
