+++ docs/2-the-zarf-cli/100-cli-commands/zarf_dev_deploy.md
@@ -1,0 +1,41 @@
+# zarf dev deploy
+<!-- Auto-generated by hack/gen-cli-docs.sh -->
+
+[beta] Creates and deploys a Zarf package from a given directory
+
+## Synopsis
+
+[beta] Creates and deploys a Zarf package from a given directory, setting options like YOLO mode for faster iteration.
+
+```
+zarf dev deploy [flags]
+```
+
+## Options
+
+```
+      --components string                  Comma-separated list of components to install.  Adding this flag will skip the init prompts for which components to install
+      --create-set stringToString          Specify package variables to set on the command line (KEY=value) (default [])
+      --deploy-set stringToString          Specify deployment variables to set on the command line (KEY=value) (default [])
+  -f, --flavor string                      The flavor of components to include in the resulting package (i.e. have a matching or empty "only.flavor" key)
+  -h, --help                               help for deploy
+      --no-yolo                            Disable the YOLO mode default override and create / deploy the package as-defined
+      --registry-override stringToString   Specify a map of domains to override on package create when pulling images (e.g. --registry-override docker.io=dockerio-reg.enterprise.intranet) (default [])
+```
+
+## Options inherited from parent commands
+
+```
+  -a, --architecture string   Architecture for OCI images and Zarf packages
+      --insecure              Allow access to insecure registries and disable other recommended security enforcements such as package checksum and signature validation. This flag should only be used if you have a specific reason and accept the reduced security posture.
+  -l, --log-level string      Log level when running Zarf. Valid options are: warn, info, debug, trace (default "info")
+      --no-color              Disable colors in output
+      --no-log-file           Disable log file creation
+      --no-progress           Disable fancy UI progress bars, spinners, logos, etc
+      --tmpdir string         Specify the temporary directory to use for intermediate files
+      --zarf-cache string     Specify the location of the Zarf cache directory (default "~/.zarf-cache")
+```
+
+## SEE ALSO
+
+* [zarf dev](zarf_dev.md)	 - Commands useful for developing packages
