+++ docs/2-the-zarf-cli/100-cli-commands/zarf_dev_find-images.md
@@ -1,18 +1,25 @@
-# zarf prepare patch-git
+# zarf dev find-images
 <!-- Auto-generated by hack/gen-cli-docs.sh -->
 
-Converts all .git URLs to the specified Zarf HOST and with the Zarf URL pattern in a given FILE.  NOTE:
-This should only be used for manifests that are not mutated by the Zarf Agent Mutating Webhook.
+Evaluates components in a Zarf file to identify images specified in their helm charts and manifests
+
+## Synopsis
+
+Evaluates components in a Zarf file to identify images specified in their helm charts and manifests.
+
+Components that have repos that host helm charts can be processed by providing the --repo-chart-path.
 
 ```
-zarf prepare patch-git HOST FILE [flags]
+zarf dev find-images [ PACKAGE ] [flags]
 ```
 
 ## Options
 
 ```
-      --git-account string   User or organization name for the git account that the repos are created under. (default "zarf-git-user")
-  -h, --help                 help for patch-git
+  -h, --help                     help for find-images
+      --kube-version string      Override the default helm template KubeVersion when performing a package chart template
+  -p, --repo-chart-path string   If git repos hold helm charts, often found with gitops tools, specify the chart path, e.g. "/" or "/chart"
+      --set stringToString       Specify package variables to set on the command line (KEY=value). Note, if using a config file, this will be set by [package.create.set]. (default [])
 ```
 
 ## Options inherited from parent commands
@@ -30,4 +37,4 @@ zarf prepare patch-git HOST FILE [flags]
 
 ## SEE ALSO
 
-* [zarf prepare](zarf_prepare.md)	 - Tools to help prepare assets for packaging
+* [zarf dev](zarf_dev.md)	 - Commands useful for developing packages
