+++ docs/2-the-zarf-cli/100-cli-commands/zarf_dev_generate-config.md
@@ -1,4 +1,4 @@
-# zarf prepare generate-config
+# zarf dev generate-config
 <!-- Auto-generated by hack/gen-cli-docs.sh -->
 
 Generates a config file for Zarf
@@ -13,7 +13,7 @@ Accepted extensions are json, toml, yaml.
 NOTE: This file must not already exist. If no filename is provided, the config will be written to the current working directory as zarf-config.toml.
 
 ```
-zarf prepare generate-config [ FILENAME ] [flags]
+zarf dev generate-config [ FILENAME ] [flags]
 ```
 
 ## Options
@@ -37,4 +37,4 @@ zarf prepare generate-config [ FILENAME ] [flags]
 
 ## SEE ALSO
 
-* [zarf prepare](zarf_prepare.md)	 - Tools to help prepare assets for packaging
+* [zarf dev](zarf_dev.md)	 - Commands useful for developing packages
