+++ docs/2-the-zarf-cli/100-cli-commands/zarf_dev_lint.md
@@ -1,17 +1,20 @@
-# zarf prepare sha256sum
+# zarf dev lint
 <!-- Auto-generated by hack/gen-cli-docs.sh -->
 
-Generates a SHA256SUM for the given file
+Verifies the package schema
+
+## Synopsis
+
+Verifies the package schema and warns the user if they have variables that won't be evaluated
 
 ```
-zarf prepare sha256sum { FILE | URL } [flags]
+zarf dev lint [ DIRECTORY ] [flags]
 ```
 
 ## Options
 
 ```
-  -e, --extract-path string   The path inside of an archive to use to calculate the sha256sum (i.e. for use with "files.extractPath")
-  -h, --help                  help for sha256sum
+  -h, --help   help for lint
 ```
 
 ## Options inherited from parent commands
@@ -29,4 +32,4 @@ zarf prepare sha256sum { FILE | URL } [flags]
 
 ## SEE ALSO
 
-* [zarf prepare](zarf_prepare.md)	 - Tools to help prepare assets for packaging
+* [zarf dev](zarf_dev.md)	 - Commands useful for developing packages
