+++ docs/2-the-zarf-cli/100-cli-commands/zarf_dev_patch-git.md
@@ -1,12 +1,18 @@
-# zarf prepare
+# zarf dev patch-git
 <!-- Auto-generated by hack/gen-cli-docs.sh -->
 
-Tools to help prepare assets for packaging
+Converts all .git URLs to the specified Zarf HOST and with the Zarf URL pattern in a given FILE.  NOTE:
+This should only be used for manifests that are not mutated by the Zarf Agent Mutating Webhook.
+
+```
+zarf dev patch-git HOST FILE [flags]
+```
 
 ## Options
 
 ```
-  -h, --help   help for prepare
+      --git-account string   User or organization name for the git account that the repos are created under. (default "zarf-git-user")
+  -h, --help                 help for patch-git
 ```
 
 ## Options inherited from parent commands
@@ -24,10 +30,4 @@ Tools to help prepare assets for packaging
 
 ## SEE ALSO
 
-* [zarf](zarf.md)	 - DevSecOps for Airgap
-* [zarf prepare find-images](zarf_prepare_find-images.md)	 - Evaluates components in a zarf file to identify images specified in their helm charts and manifests
-* [zarf prepare generate-config](zarf_prepare_generate-config.md)	 - Generates a config file for Zarf
-* [zarf prepare lint](zarf_prepare_lint.md)	 - Verifies the package schema
-* [zarf prepare patch-git](zarf_prepare_patch-git.md)	 - Converts all .git URLs to the specified Zarf HOST and with the Zarf URL pattern in a given FILE.  NOTE:
-This should only be used for manifests that are not mutated by the Zarf Agent Mutating Webhook.
-* [zarf prepare sha256sum](zarf_prepare_sha256sum.md)	 - Generates a SHA256SUM for the given file
+* [zarf dev](zarf_dev.md)	 - Commands useful for developing packages
