+++ docs/2-the-zarf-cli/100-cli-commands/zarf_dev_sha256sum.md
@@ -1,25 +1,17 @@
-# zarf prepare find-images
+# zarf dev sha256sum
 <!-- Auto-generated by hack/gen-cli-docs.sh -->
 
-Evaluates components in a zarf file to identify images specified in their helm charts and manifests
-
-## Synopsis
-
-Evaluates components in a zarf file to identify images specified in their helm charts and manifests.
-
-Components that have repos that host helm charts can be processed by providing the --repo-chart-path.
+Generates a SHA256SUM for the given file
 
 ```
-zarf prepare find-images [ PACKAGE ] [flags]
+zarf dev sha256sum { FILE | URL } [flags]
 ```
 
 ## Options
 
 ```
-  -h, --help                     help for find-images
-      --kube-version string      Override the default helm template KubeVersion when performing a package chart template
-  -p, --repo-chart-path string   If git repos hold helm charts, often found with gitops tools, specify the chart path, e.g. "/" or "/chart"
-      --set stringToString       Specify package variables to set on the command line (KEY=value). Note, if using a config file, this will be set by [package.create.set]. (default [])
+  -e, --extract-path string   The path inside of an archive to use to calculate the sha256sum (i.e. for use with "files.extractPath")
+  -h, --help                  help for sha256sum
 ```
 
 ## Options inherited from parent commands
@@ -37,4 +29,4 @@ zarf prepare find-images [ PACKAGE ] [flags]
 
 ## SEE ALSO
 
-* [zarf prepare](zarf_prepare.md)	 - Tools to help prepare assets for packaging
+* [zarf dev](zarf_dev.md)	 - Commands useful for developing packages
